Add page title to AlbumProfile and refactor Photos component
