+++ client/src/data/images.ts
@@ -1,0 +1,38 @@
+type ImageItem = {
+	imageUrl: string;
+	albumName: string;
+	photoName: string;
+};
+
+export const images: ImageItem[] = [
+	{
+		imageUrl: "https://pagedone.io/asset/uploads/1688031162.jpg",
+		albumName: "Vintage Cars",
+		photoName: "Mercedes Benz",
+	},
+	{
+		imageUrl: "https://pagedone.io/asset/uploads/1688031232.jpg",
+		albumName: "Vintage Cars",
+		photoName: "BMW",
+	},
+	{
+		imageUrl: "https://pagedone.io/asset/uploads/1688031357.jpg",
+		albumName: "Vintage Cars",
+		photoName: "Audi",
+	},
+	{
+		imageUrl: "https://pagedone.io/asset/uploads/1688031375.jpg",
+		albumName: "Vintage Cars",
+		photoName: "Porsche",
+	},
+	{
+		imageUrl: "https://pagedone.io/asset/uploads/1688031396.jpg",
+		albumName: "Vintage Cars",
+		photoName: "Ferrari",
+	},
+	{
+		imageUrl: "https://pagedone.io/asset/uploads/1688031414.png",
+		albumName: "Vintage Cars",
+		photoName: "Lamborghini",
+	},
+];
